Do not raise an error when a framework reregisters using a new user
Currently, it is not possible to restart an existing framework using a
new user without draining mesos agents hosting its tasks + triggering a
mesos leader election.
This is not acceptable in production.

This commit transforms the validation error into a simple warning.
This is not an issue on our infrastructure, since we always specify a
user for tasks launched by frameworks.

JIRA: MESOS-4868